+++ lib/sendemail-gmail.js
@@ -1,8 +1,14 @@
+var google       = require('googleapis');
+var gmail        = google.gmail('v1');
+var OAuth2       = google.auth.OAuth2;
+var oauth2Client = new OAuth2(process.env.GOOGLE_CLIENT_ID, process.env.GOOGLE_CLIENT_SECRET);
+var btoa         = require('btoa'); // encode email string to base64
+
 /**
  * sendEmail abstracts the complexity of sending an email via the GMail API
- * @param {Object} options - the options for your email, these include:
+ * @param {Object} request - the options for your email, these include:
  *  - credentials
- *    - {Object} tokens - the list of tokens returned after Google OAuth
+ *    - {Object} auth - the list of tokens returned after Google OAuth
  *    - {Array} emails - the current user's email addresses (List)
  *  - {String} to - the recipient of the email
  *  - {String} from - sender address
@@ -12,6 +18,25 @@
  *   @arg {Object} error - the error returned by the GMail API
  *   @arg {Object} response - response sent by GMail API
  */
-module.exports = function sendEmail(options, callback) {
+module.exports = function sendEmail(request, callback) {
+  var credentials = request.auth.credentials;
+  oauth2Client.setCredentials(credentials.tokens);
+  var params = { userId: 'me', auth: oauth2Client, resource: {} };
 
+  var email  = request.auth.credentials.emails[0].value;
+  var base64EncodedEmail = btoa(
+        "Content-Type:  text/html; charset=\"UTF-8\"\n" +
+        "Content-length: 5000\n" +
+        "Content-Transfer-Encoding: message/rfc2822\n" +
+        "to: nodecoder@gmail.com\n" +
+        // "from: \"test\ <"+ email +">\n" +
+         "from: nodecoder@gmail.com \n" +
+        "subject: Tell Me when you get this!!\n\n" +
+
+        "This message was sent by <b>Node.js</b>!"
+          ).replace(/\+/g, '-').replace(/\//g, '_');
+  var params = { userId: 'me', auth: oauth2Client, resource: {} };
+  // see: http://stackoverflow.com/questions/30590988/failed-sending-mail-through-google-api-with-javascript
+  params.resource.raw = base64EncodedEmail;
+  return gmail.users.messages.send(params, callback);
 }
